a bit of error checking
git-svn-id: svn+ssh://svn.sw1v.org/svn/sys/trunk/amazon@945 d43e5919-7b50-4614-a95a-1017a2390aba

+++ bin/stop-instance.sh
@@ -5,6 +5,13 @@ set -e
 toolsdir=$(cd `dirname "$0"` && pwd)
 wd="$1"
 instance_id=`cat "$wd/instance_id"`
-"${toolsdir}/aws" terminate-instances "$instance_id" > /dev/null
+
+out=`"${toolsdir}/aws" --xml terminate-instances "$instance_id"`
+if echo $out | grep -i '<error>' > /dev/null; then
+    echo -e "error terminating instance:" >&2
+    echo $out >&2
+    exit 1
+fi
+
 rm -r "$wd"
